Move LockManager API back to using int timeouts, since that's the way they're handled elsewhere (FetchConfiguration, Broker, etc). Only use the lock timeout on a forUpdate query if it is greater than the configured query timeout. Selects that are made *only* to lock a row (rather than to query data, with locking as a side effect) still use the lock timeout exclusively, ignoring the query timeout.
git-svn-id: https://svn.apache.org/repos/asf/incubator/openjpa/trunk@526192 13f79535-47bb-0310-9956-ffa450edef68

+++ openjpa-kernel/src/main/java/org/apache/openjpa/kernel/VersionLockManager.java
@@ -90,7 +90,7 @@ public class VersionLockManager
      *
      * @see StoreContext#transactional
      */
-    protected void lockInternal(OpenJPAStateManager sm, int level, long timeout,
+    protected void lockInternal(OpenJPAStateManager sm, int level, int timeout,
         Object sdata) {
         // Set lock level first to prevent infinite recursion with
         // transactional(..) call
